♻️ [#174] Refactor: ScheduleInput 파일 분리
